+++ src/main/java/micdoodle8/mods/galacticraft/core/entities/player/PlayerClient.java
@@ -24,7 +24,11 @@ import micdoodle8.mods.galacticraft.core.wrappers.PlayerGearData;
 import net.minecraft.client.audio.PositionedSoundRecord;
 import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.util.*;
+import net.minecraft.util.ChatComponentText;
+import net.minecraft.util.ChunkCoordinates;
+import net.minecraft.util.IChatComponent;
+import net.minecraft.util.MathHelper;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.ChunkCoordIntPair;
 import net.minecraftforge.common.MinecraftForge;
 
@@ -209,7 +213,7 @@ public class PlayerClient implements IPlayerClient
                 {
                 	player.height = 1.8F;
                 }
-            	player.boundingBox.maxY = player.boundingBox.minY + (double)player.height;
+            	player.boundingBox.maxY = player.boundingBox.minY + player.height;
             }
         }
